requirements.txt
python-telegram-bot
telethon
selenium
python-dotenv